+++ RPGHeroes/Items/Weapon.cs
@@ -19,10 +19,10 @@ namespace RPGHeroes.Items
     }
     public class Weapon : Item
     {
-        public int WeaponDamage { get; set; }
+        public double WeaponDamage { get; set; }
         public WeaponType Type { get; set; }
 
-        public Weapon(string name, int requiredLevel, Slot itemSlot, WeaponType type, int weaponDamage) : base(name, requiredLevel, itemSlot)
+        public Weapon(string name, int requiredLevel, Slot itemSlot, WeaponType type, double weaponDamage) : base(name, requiredLevel, itemSlot)
         {
             Type = type;
             WeaponDamage = weaponDamage;
